EID-1328 Rename EidasSamlParserService to EidasSamlParserProxy
This is to remain consistent with the VerifyServiceProviderProxy in
shared.
